image methods to arprocimg
